chore: replace usages of TypeToken with geantyref (#1133)
### Motivation
The TypeToken implementation of guava/gson has the downside of always
needing to create 2 instances: The TypeToken itself and the Type that is
actually captured by it. Furthermore, the old bundled version of gson in
minecraft makes it harder to use newer features of TypeToken, especially
`TypeToken.getParameterized`.

### Modification
Replace all relevant usages of TypeToken with geantyref (which is a
generic type reflection library). This eliminates the need for usages of
TypeToken and has some great util methods when it comes to working with
generic types.

### Result
No more usages of TypeToken that are replaceable with geantyref.

##### Other context
Resolves #1036