chore: unify document properties (#1138)
### Motivation
At the moment there are different types of document properties. A
DocProperty which can be appended and read from everything that
implements DocPropertyHolder and a ServiceProperty which is exclusive to
services. This means that we're basically maintaining two different
document property apis, with the service property one having some more
methods that were missing in the DocProperty api.

### Modification
Removed the service property implementation completely and improved the
doc property api to allow for everything that the current service
property api allowed. Furthermore, the api was expanded and better
integrated into the cloudnet system, to match the design of existing api
classes.

### Result
One general and now unified api to read and write doc properties.

##### Other context
Fixes #1072