chore: add better equals/hashCode implementation for NetworkClusterNode (#1142)
### Motivation
The current way lombok implements equals and hashCode is to include all
fields. This is incorrect as a network nodes are identified by the
unique id, not by unique id, listeners and properties.

### Modification
Add a custom hashCode and equals method to NetworkClusterNode that only
takes the unique id into account.

### Result
NetworkClusterNode instances are now only identified by the unique id.