fix: properly order console tab suggestions (#1152)
### Motivation
The current candidate matching system in CloudNet is using a jaro
winkler distance to check if a possible suggestion matches the given
started input and sorts the values based on the algorithm result. That
ordering is not properly kept after passing the candidates to jline.

### Modification
Set the sort of the candidates that are passed to jline. This will keep
the candidates in the order given from the called tab complete handler.

### Result
JLine now displays the tab completions for a command in the intended
order.

##### Other context
Before:

![image](https://user-images.githubusercontent.com/40468651/222922696-89dc050d-d5a3-4aff-b61d-ecf4dca913d6.png)

After:

![image](https://user-images.githubusercontent.com/40468651/222922700-f7efe616-3f67-45b5-8b04-51bc7dd5d006.png)