fix: properly close buffers to prevent memory leaks (#1146)
### Motivation
The current way we handle data bufs is prone to memory "leaks". They are
not noticeable by the user directly, as the underlying memory of a
buffer gets cleaned when it becomes phantom-reachable, but if a process
is allowed to allocate a lot of memory, the consumed memory will rise
until the garbage collector takes action (which can be very late
considering that we set ZGC as the default in our start scripts).

### Modification
Change the complete internal code in a way that no more memory leaks are
occurring. Note that this change might be error prone, as external
plugins might use buffers in an invalid way that causes issues after the
change. External code is now required to acquire and release a buffer to
ensure it's not getting closed, rather than keeping a reference which
prevents the buffer from getting released.

### Result
No more internal memory leaks.