fix: don't add the initial task template to the group configuration as well (#1175)
### Motivation
We found in the group configuration that it contains the template of the
tasks (created on the initial setup). I asked on the Discord if this
were a bug, because the template were already on the task configuration
and that it shouldn't be there was confirmed by: 0utplay#0187 (https://discord.com/channels/325362837184577536/818777626663321671/1089277611378753597)

### Modification
I've modified the DefaultTaskSetup class to not include the template of
the task to the group

### Result
Now the template of the tasks (created on the initial setup) will not be
inside the group configuration.