fix: ignore http requests with an opaque uri (#1181)
### Motivation
Opaque URIs are a type of URI that does not allow access to the
underlying resource. Instead, the opaque URI contains a string of
characters that represent the resource, but the handler has no
information about how to access or manipulate that resource. As our http
handler design is based on the assumption that the path to the resource
to call is present, we cannot handle requests that are opaque at all.

### Modification
Early deny requests that are sending an opaque request and respond with
status 400 (bad request) to them.

### Result
There are no more requests processed that cannot get handled and are
causing exceptions.