chore: rewrite document api (#1178)
### Motivation
At the moment the document api is very limited due to the reason that
practically only the json implementation is used everywhere. While there
is an abstract interface for a Document that specifies most of the
methods implemented in the json implementation, that interface cannot
really be used for custom implementations as the exposed api never
accepts the custom implementation.

### Modification
The whole Document api has been rewritten and the Document interface is
now used everywhere in the api, allowing to use custom document types.
To make this work, this PR also contains a way to transfer the key-value
pairs from one document type to another. To follow the internal
structure of having one mutable type and one immutable type for
everything (to make it clear when something can/should be edited, an
when that shouldn't be the case) the document is now split up into
mutable and immutable as well.

### Result
Way cleaner document implementation with proper support for custom
document types.