fix: return empty document for non-existent file when parsing from path (#1197)
### Motivation
The document api rewrite introduced a breaking change (which affects
internals as well) as the `DocumentFactory#read(Path)` (prior
`Document#read(Path)`) now throws an exception in case the file does not
exist rather than returing a new, empty document.

### Modification
Let `DocumentFactory#read(Path)` return a new, empty document in case
the file does not exist or is not a file and change the specification of
the method to reflect the change.

### Result
The `DocumentFactory#read(Path)` method now returns an empty document in
case the file does either not exist or is not a regular file.