[ci skip] ci: make build action a requirement for merge group (#1203)
### Motivation
To enable merge groups we need an action that reports if a change can be
merged into the target branch.

### Modification
Trigger the build action on the `merge_group` dispatch as well. This
tells GitHub that the action reports the status check.

### Result
Merge groups have a way to figure out if a change can be merged into the
target branch.