fix: remove usages of checkerframework (#1209)
### Motivation
There are two accidental usages of checkerframework in classes that are
always handled by the compile cache for incremental builds. Therefore
when executing a fully clean build it will fail because of these two
imports.

### Modification
Remove usages of checkerframework.

### Result
No more usages of checkerframework and first time builds pass again.