ci: update action configuration (#1214)
### Motivation
The current action configuration does no longer properly keep up with
our requirements.

### Modification
A list of all modifications:
* There are no longer caches stored for each branch. While the action
will be able to restore a cache, it's always based on nightly and only
the nightly branch will update the cache. This should reduce the amount
of stored caches.
* The `gradle-build-action` is now used to manage the caches. The action
manages the cache way better than the current, default configuration for
gradle caches shipped with the `setup-java` action.
* A specified action step was added that just caches the created Fabric
Loom cache. The cache restore is based on the dependency versions and
gradle version (it doesn't seems like there is a better way to cache the
loom output rather than based off on all dependencies).
* An action step was added that attaches the test results to pull
requests and the workflow run.
* Both CodeQL and the Maven Central publish action now use the read-only
gradle cache as well.
* The build process is no longer executed on Java 19 alongside Java 17.
This will be re-enabled when Kotlin finally supports Java 20 or 21, and
it actually makes sense again to test.

### Result
Faster build times and a better action configuration.