fix: log throwable catched when ticking signs (#1219)
### Motivation
We catch all throwables while ticking a sign. If any exception occurs we
log a warning but that warning is missing the catched throwable and
therefore we don't have any information about the reason for the error.

### Modification
Made sure to print the throwable too.

### Result
We have information if something goes wrong