chore: update openapi specs due to recent ServiceConfiguration changes (#1218)
### Motivation
The OpenAPI specs had to be adapted due to the past changes to the
service configuration. In addition, failure to create a service would
lead to an exception on the node, as the ServiceCreationResult could not
be serialized.

### Modification
Adopted the changes in the OpenAPI specs and made the
ServiceCreationResult a class resolving the serialization errors.

### Result
The OpenAPI specs are up-to date and the serialization is working
properly