feat: warn about missing manifests and main classes (#1217)
### Motivation
If a application does not contain a manifest reading the attributes will
fail as the manifest is null. Same goes for the Main-Class, if the
manifest does not contain a main class the main class is passed down
into the process builder as null which will result in an exception that
does not say anything about the main class as it furtherdown the
execution chain.

### Modification
Added a new method `preCheckJarManifest` that takes a nullable manifest
and ensures that a manifest is present and the manifest contains a main
class. Otherwise two exceptions are thrown for each case.

### Result
It is clear what is the issue with the application jar and the service
is deleted afterwards instead of staying prepared.