feat: enable injection of services from the service registry (#1232)
### Motivation
We wanted to replace the service registry fully with injection but that
would result in a weird injection tree and instead we can just make it
easier to use the service registry when using injection.

### Modification
Added a new `@Service` annotation which retrieves the declared element
from the service registry instead of any injection binding.
 
### Result
Injection and the service registry go hand in hand and there is no need
to remove the service registry