chore: replace service registry usages with service inject annotation (#1234)
### Motivation
#1232 introduced a new `@Service` annotation that allows to injection
providers from the service registry directly instead of injecting the
service registry and retrieving it from the registry afterwards. This PR
makes use of the new annotation and replaces the service registry where
possible.

### Modification
Deprecated the ServiceRegistry.first(Class) method as it requires
injection abuse to work. Replaced other usages of the service registry
with the new annotation where possible.

### Result
We are using the new Service annotation.