chore: bump npc lib to beta5, fix some smaller npc issues observed during testing (#1257)
### Motivation
Npc-Lib beta5 released which fixes a problem for all servers that can't
use a modern variant to load profiles (like the Paper PlayerProfile
api). In addition, the new version had a breaking change in how the
parameters for the method related to entitiy metadata are sorted.

### Modification
Bump the npc-lib to beta5 and fix the parameter order of the previously
mentioned methods. This PR also introduces a fix for the following two
issues (in the CloudNet code):
* When an npc was spawned to a player, the packages needed to update the
entity meta and inventory item packets were sent to all tracked players,
not just the new player (who is also the only player that needs the
packets).
* If the "usePlayerSkin" option was used for an npc, the npc no longer
spawned because the npc profile had the same name and uuid as the online
player.

### Result
The npc lib has been updated and some minor issues with the npcs have
been fixed.