fix: draining causes jvm halt instead of a stop (#1262)
### Motivation
When enabling draining on a node and stopping all already running
services the node should be stopped. Currently instead of stopping the
node we cause the jvm to halt therefore the node does not stop correctly
and does not exit at all.

### Modification
Used the shutdown method in the shutdown handler to stop the node and
exit.

### Result
The node exits successfully.