fix: the launchers memory should not effect the node memory (#1221)
### Motivation
The way our launcher works is that all jvm options of the launcher get
passed down to the started node instance. Therefore increasing the
memory in one of the start files would result in the memory increasing
for two different processes.
1. The launcher process it self start is started by the start file
2. The memory of the new node process started by the launcher

So that means that the memory consumption is doubled.

### Modification
Made sure to remove the -xmx and -xms options of the launcher before
passing the remaining options to the node and introduced a new option in
the launcher.cnl.
The new option `cloudnet.node.memory` is the new way to set the memory
of the node (in megabytes).

### Result
The memory is not occupied twice and the node memory is not dependant on
the launcher.