fix: correctly ignore EOT signals (#1278)
### Motivation
EOT is a control character that signals "end of text" or "end of
transmission". This behaviour can be triggered by typing CTRL+D into the
console. At the moment we're already handling that by returning null
from the `readLine` method, which leads to an exit of the console-read
while loop due to a null check (and therefore the terminal is no longer
useable after it).

### Modification
Remove the extra `readLine` method and move the line reading and error
handling directly to the while loop. This allows to just silently ignore
EOT input without breaking the while loop, and allows for future error
handling too (without having to introduce some sort of error return
handling stuff to `readLine` in order to find out which kind of error
occurred while reading the line input).

### Result
EOT signals are now handled correctly and are silently ignored by the
console reading loop.