feat: add object serializers for all kinds of time types (#1285)
### Motivation
There were multiple people in the past few days that tried to serialize
a java.time class using the default object mapping system. At the moment
this is not possible due to the lack of serializers for these types.

### Modification
Add serializers for all java.time classes that are made to support the
ISO-8601 standard. Other calendar systems (such as the Chinese one) are
still not supported.

### Result
Working with RPC and ISO-8601 java time types is now fully supported out
of the box and no custom serializers are required anymore.