fix: make sure to create parent directories in sftp (#1225)
### Motivation
When creating a new file on a sftp storage in a non existant directory
the file creation will fail instead of creating the parent directories.
This will result in an exception when installing a template on a sftp
storage due to the creation of the config/sponge/sponge.conf file.

### Modification
Made sure to create the directories before opening an outputstream or
creating a file in a sftp template storage. Furthermore the tests were
extended to cover the creation of files in non existant directories

### Result
The parent directories are created.