fix: ensure that bridge is loaded before property access (#1317)
### Motivation
The method to fill common placeholders is used for cloudnet signs too
and as the service is starting the signs are trying to resolve some of
the bridge properties. That fails because the service is just starting
and the bridge is not loaded yet.

### Modification
Added a check that makes sure that we are only replacing placeholders if
they are present.

### Result
Signs are working correctly again.