fix: ensure that npcs are removed correctly (#1349)
### Motivation
When creating a npc on a service created from a static task, the
infolines (and for entity npcs even the entity itself) are saved to the
disk. After restarting the service the entity and info lines are
duplicated because the plugin respawns them (as it should) but there are
still the old ones from the disk.

### Modification
Added a new close method to the PlatformNPCManagement. By default it
removes all tracked entities (thus all info lines are removed too). The
new method is called on plugin disable.

### Result
No duplicates.

##### Other context
Fixes #1348