feat: ignore non-existent worlds when choosing a free sign (#1371)
### Motivation
If multiple tasks are in the same group and the tasks use different
world names, it is possible that the CloudSigns module selects signs
that do not exist, therefore the services are never displayed. This PR
ignores all signs that do not exist on the server.

### Modification
Signs in unknown worlds are skipped.
Another "cleanupall" option has been added to the cloud signs command,
which behaves like the old "cleanup".
The old "cleanup" option has been extended by an additional optional
argument "world", which specifies the world in which the cleanup should
take place.

### Result
Only characters in existing worlds are considered.

##### Other context

https://discord.com/channels/325362837184577536/818777626663321671/1215677639709425756