fix: append app file to class path property before launching application (#1411)
### Motivation
Due to the fact that we use the wrapper to start the application the jvm
misses to add the platform jar to the classpath. But we also can't add
the platform to the classpath before the wrapper is not started as this
would cause us to start the platform directly.

### Modification
After the wrapper is started we append the platform jar to the class
path property. Therefore the jar is not loaded directly but is still
part of the class path property and platforms can depend on the
property.

### Result
Platforms can depend on the default java behaviour which is to add the
own platform to the cp.

##### Other context

https://discord.com/channels/325362837184577536/818777626663321671/1243214044853309483)