chore: use jline ffm terminal implementation (#1441)
### Motivation
JLine includes a FFM terminal implementation, which became a permantent
feature in JDK 22. As we're now requring java 22, we can use the FFM
terminal instead of some jni-based implementation.

### Modification
* Replaced the usages of `org.jline:jline`: that is a complete bundle
containing all jline dependencies, which is not nessacary at all. We're
now including only the line reader and terminal dependencies.
* Enable native access: while java itself has no requirement for this
*yet*, some steps were already made into that direction coming to java
24 (see JEP 472). JLine has an explicit check that verifies that native
access is enabled, so setting the flag is mandatory in order to use the
FFM terminal implementation.
* Replaced usages of `AnsiConsole.systemInstall/systemUninstall`: the
ansi handling is already baked into JLine itself, these methods are only
overriding the systems streams nowadays. No native lib is required for
ansi anymore, so the color support detection becomes useless as well. If
I understood the code correctly, ansi color codes are stripped from the
output by JLine if it detects a terminal without color support.

This change was tested on Debian 12 and Windows using cmd and PowerShell
5 & 7. All animations, tab completion & typing works as expected.

### Result
We're now using the FFM-based terminal implementaton of JLine and are no
longer relying on JNI libraries to do the job.