fix: prevent blocking operations on the netty event loop (#1461)
### Motivation
Currently it's possible to deadlock the netty event loop, This is caused
by the fact that the network task scheduler will now execute packets
that cannot be queued (when the queue is full) on the event loop. The
user might however call another blocking network operation on the event
loop (like sending a query packet) which means that the incoming packets
(like a response to the query) cannot be processed and therefore the
event loop is deadlocked.

### Modification
The event loop is never used as a fallback to call network operations
anymore. Instead, a new second thread pool was added that can add
infinite threads in case the network task queue gets fuller and fuller.
The threads are started with an initial delay of 150ms which is stadely
decreased to 50ms in case the network task queue gets even fuller over
time. This is to prevent starting useless threads when just waiting a
tiny bit for another thread to finish work would've been suffiecent.
Idle threads in the fallback thread pool are removed after 15 seconds of
idling.

### Result
The netty event loop doesn't execute blocking operations anymore (which
prevents all incoming packets from being handled)

##### Other context
Fixes #1457