drop: packet receive and packet send event (#1464)
### Motivation
The packet send and receive event can be dangerous in three situations:
1. both events can be triggered from an event loop. An unaware user
might issue a blocking operations that will cause the event loop to
stall.
2. the event has access to the packets being sent and received and could
modify them. If done wrong, this could break further packet handlers
(not only locally but also on remote network components).
3. if a query packet send or response is cancelled it might cause
another component to stall forever if it blocks on the packet response.

Another possible solution would be to just make the packet not
cancellable anymore, but this would still have the risk of point 1 & 2.

### Modification
Drops the `NetworkChannelPacketReceiveEvent` &
`NetworkChannelPacketSendEvent` and removes all usages of it. Also
rewrites the `NettyNetworkChannel.sendPacket(Packet...)` method as it is
no longer a problem to execute the send operation on the event loop.

### Result
The `NetworkChannelPacketReceiveEvent` & `NetworkChannelPacketSendEvent`
no longer exist and are not called anymore.