feat: Run post-initialization method at the first server tick
### Motivation

Certain features of CloudNet react to services updating and thus changing their state to "Running". In the current implementation, this update happens at the time the bridge module/plugin/extension is enabled.
This timing is not ideal as a lot of time-consuming tasks, like generating or loading the worlds/dimensions and other modules/plugins/extensions may happen after that.
This causes the other components, like the signs module or the NPCs module, to report those services as being already online and ready to join by players, whilst the opposite is actually the case.

### Modification

All currently supported server software platform integrations are augmented, so that `PlatformBridgeManagement#postInit` is called in the first game tick. This tick is only called, once the server is actually initialized, loaded and running. This also means that the server is ready to accept players.

### Result

Amongst other things, signs now show the server for joining only, once the server is actually ready to accept players.