feat: add caching policies for remote inclusions (#1413)
### Motivation
Some users use remote inclusions to always download latest files of
certain plugins. The automatic caches makes that a bit hard.

### Modification
Added a new `cacheStrategy` option for remote inclusions and we cache
files according to the provided strategy. Only two strategies are supported
for now, however support for further customizing will be added in a future
step.

### Result
The caching can be customized as needed.