chore: update GitHub actions (#1474)
### Motivation
Some of our GitHub actions are outdated but these updates were either
not detected by renovate or cannot be done automatically.

### Modification
- Bump `Vankka/pr-target-branch-action` to v3: this resolves a warning
because the old 2.1 version was still running on node16 which is now
deprecated. Additionally a mistake was fixed: only beta should be
allowed to PR into release in the end.
- Fix a deprecation warning of `github/codeql-action/init`: The `latest`
tool option name was changed to `linked` to reflect the actual CodeQL
version being used: the version the action was linked (build) with.
- Buimp `gradle/actions/setup-gradle` to v4: The wrapper validation step
can be removed as this is now part of the action itself. Additionally I
enabled that the job summary is always appended to all GitHub action
runs.

### Result
The GitHub actions versions used in the project are up to date and all
relevant deprecations are resolved.

##### Other context
Supersedes #1472