chore: replace deprecated converterClass property in logback config (#1491)
### Motivation
Logback introduced a new deprecation in the last release which wasn't
mentioned in the changelog. This a huge logback debug output when
starting CloudNet to be printed into the console. Essentially the
`converterClass` property of `conversionRule` was renamed to `class`.

Deprecation notice: `|-WARN in
ch.qos.logback.core.joran.action.ConversionRuleAction - [converterClass]
attribute is deprecated and replaced by [class]. See element
[conversionRule] near line 29`

### Modification
Replace `converterClass` attribute of `conversionRule` with `class` to
resolve the deprecation notice.

### Result
The deprecation is resolved and no more debug logs are printed when
starting CloudNet.