fix: testcontainers related test issues (#1492)
### Motivation
At the moment tests cannot be execurted locally due to test failures
related to MariaDB. For some reason, the mariadb server issues a SSL
certificate which becomes valid after a few seconds, however, the
connection to the server is established 1 or 2 seconds before that. This
leads to an exception being thrown as the certificate is not yet valid.

Furthermore, due to enabled debug logging we get debug logs containing
stack traces from docker-java which are interpreted as test failures by
IJ (while the build completes normally).

### Modification
The MariaDB issue can be fixed by simply disabling SSL in the mariadb
container. The logging issue is fixed by applying the recommanded
logging configuration of testcontainers (see
https://java.testcontainers.org/supported_docker_environment/logging_config)
which disables the debug logging of docker-java. Therefore the
misinterpreted errors are no longer printed into the console.

### Result
Tests work locally as expected again.