fix: use high timeout for query packets instead of weak values (#1500)
### Motivation
The new implementation of query packets resulted in possible dead locks. The dead locks were caused, when the QueryPacketManager lost the reference to the stored future, because the answer to the query would never be handled as the entry in the map is lost.

### Modification
Removed the weakValues option from the backing cache and use a really high timeout of 8 hours for cache invalidation. Usually the cache should be empty so the high timeout value is not a problem.

### Result
No more deadlocks with query packets.

##### Other context
Fixes #1480