fix: remove dev command translations & sync german translations (#1528)
### Motivation
The crowdin automation has done some updates to our translations. We do
not want to pull all changes from crowdin as some languages are poorly
translated and not complete yet.

### Modification
Removed the usage of translations from the dev command as this command
should rely on as little as possible.
Furthermore the changes from crowdin related to the german translations
were pulled.

### Result
Correct german translations and no translations in the dev command
anymore.