fix: add the default task setup back to the initial setup (#1534)
### Motivation
After the removal of CloudPerms the initial setup did not contain the
default task setup for `Lobby` and `Proxy` anymore because it only
relied on the fact that the task provider was initialized before the
setup was started. The removal of CloudPerms also removed that early
init and thus the setup was not complete anymore.

### Modification
Added the task provider as parameter to the setup step of the node.
Aerogel is forced to init the provider first and therefore the setup is
registered before the actual start of the setup.

### Result
The default task setup is back.