fix: commodore class remap on latest spigot versions (#1535)
### Motivation
Spigot made a change to Commodore a few days ago (merged into paper
yesterday) that made the `Commodore#convert` non-static. Therefore our
current transformer for the class fails as it tries to return the slot 0
(which, when the method is static, is the first argument) which now is
the Commodore instance.

### Modification
Check if the method being transformed is static and return slot 0 in
that case, if the method is not static return slot 1. In both
constellations the first argument of the convert method, the raw input
class bytes, is returned.

### Result
Latest Spigot/Paper Services no longer crash due to a bad stack frame.